Split E2E tests into per-topic files
